HW3

+++ LoanCalc.java
@@ -86,7 +86,7 @@ public class LoanCalc {
                 break;
             }
         }
-        return payment;  // Return the final payment value
+        return (int) payment;  // Return the final payment value
     }
 
     // Uses bisection search to compute an approximation of the periodical payment
@@ -113,6 +113,6 @@ public class LoanCalc {
             }
         }
 
-        return (low + high) / 2; // Return the approximate payment
+        return (int)((low + high) / 2); // Return the approximate payment
     }
 }
